requirements.txt
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0          
art==6.2              
attrs==24.2.0         
blinker==1.8.2        
cachetools==5.4.0     
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lapx==0.5.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
narwhals==1.4.1
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.27.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.1
rich==13.7.1
rpds-py==0.20.0
scipy==1.13.0
seaborn==0.13.2
setuptools==72.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.37.1
sympy==1.12
tbb==2021.12.0
tenacity==8.5.0
thop==0.1.1.post2209072238
toml==0.10.2
torch==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.18
urllib3==2.2.1
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.23.0
websockets==12.0
